PDF POEMS COMPLETE
LaTeX is harder than I thought it would be but this is pretty close
to complete. Maybe someday I can refine it further, but I'm calling
this done for now. The results look pretty good, actually. Not perfect
but this will have to do.

Maye someday the headers can be redrawn?